pyproject.toml
[tool.poetry]
name = "qonto-to-odoo"
version = "0.1.0"
description = "App to connect Qonto and Odoo"
authors = ["Jonas TURBEAUX"]
license = "GPL3"
readme = "README.md"
packages = [{include = "qonto_to_odoo"}]

[tool.poetry.dependencies]
python = "^3.10"
jupyter = "^1.0.0"
Flask = "^2.2.3"
requests = "^2.28.2"
ipython = "^8.11.0"
ipdb = "^0.13.11"
jupyterthemes = "^0.20.0"
django = "^4.1.7"
django-solo = "^2.0.0"
django-extensions = "^3.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
